requirements.txt
Django
djangorestframework
django-cors-headers
PyMySQL
openai
python-dotenv
requests
